Added Hyper Flux Version of the Workflow
u/shootthesound on Reddit shared some improvements for performance on cards with lower specs, this modified version of the Flux Latent Upscaler workflow uses the new text encoder model which is said to improve details in Flux: ViT-L-14-TEXT-detail-improved-hiT-GmP-TE-only-HF

https://huggingface.co/zer0int/CLIP-GmP-ViT-L-14/blob/main/ViT-L-14-TEXT-detail-improved-hiT-GmP-TE-only-HF.safetensors

Place it in models/clip

Bytedance 16 step and 8 Step Flux loras are used to cut steps and speed up inference

https://huggingface.co/ByteDance/Hyper-SD/blob/main/Hyper-FLUX.1-dev-16steps-lora.safetensors

https://huggingface.co/ByteDance/Hyper-SD/blob/main/Hyper-FLUX.1-dev-8steps-lora.safetensors

Place them both in models/loras

u/shootthesound also adjusted settings on Model Sampling Flux node 

Step count changes at all stages cut the time from ~280 seconds to just over 100 seconds using a 4090 with 24GB of VRAM, the speed improvements should be noticeable with all cards capable of running Flux dev. There is a slight degradation in quality but the reduction in system resources and time makes it worth it, especially if running lower on specs. Some changes occur in the final output but overall this is very close to the original workflow in terms of detail and prompt adherence, of course this is a different approach and results may vary.